backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
pyyaml==6.0.2
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
openai==1.58.1
httpx==0.28.1
aiosqlite==0.20.0
python-multipart==0.0.20
websockets==14.1
